+++ .eslint-doc-generatorrc.js
@@ -1,0 +1,11 @@
+/** @type {import('eslint-doc-generator').GenerateOptions} */
+const config = {
+  configEmoji: [
+    ['jsx-runtime', '🏃'],
+    ['recommended', '☑️'],
+  ],
+  ignoreConfig: ['all'],
+  urlConfigs: 'https://github.com/jsx-eslint/eslint-plugin-react/#shareable-configs',
+};
+
+module.exports = config;